Update 20230311201746_setup-vector-store.sql
This adds the ID value of the row the document belongs to into the metadata so it is fetched as part of every element in the  Document[] array.  This is because, out of the box. Document object only brings in:     
pageContent: string;
metadata: Metadata; 

I found myself needing to keep track of these records in the database for my needs and this changed helped me avoid multiple calls to DB.

The same change can be added to the kw_match_documents_2 as well.